Check for the moderate option being on before sending pending email
We noticed that this email gets sent regardless if the moderate new users option is set or not.  We had a client that ended up only wanting the feature of restricting BuddyPress pages to registered users, but decided to not use the moderate feature and users were still getting pending user emails after activation even though moderate feature was off.  This change just creates a check and only utilizes the functionality if the feature is active.